+++ attributes/default.rb
@@ -1,8 +1,8 @@
 # default recipe
 default['chef-backup']['log_dir'] = '/var/log/chef-backup'
 default['chef-backup']['backup_dir'] = '/opt/chef-backup'
-# system user
 default['chef-backup']['backup_user'] = 'chef-backup'
+
 # knife.rb config
 default['chef-backup']['knife']['client_name'] = 'chef-backup'
 default['chef-backup']['knife']['chef_server_url'] = 'https://localhost'
@@ -20,3 +20,10 @@ default['chef-backup']['restore_file'] = 'chef.backup.latest.tar.gz' # expected 
 default['chef-backup']['restore_cron']['hour'] = '1'
 default['chef-backup']['restore_cron']['minute'] = '0'
 default['chef-backup']['restore_cron']['weekday'] = '*'
+
+# s3 sync recipe
+default['chef-backup']['s3sync']['bucket'] = nil
+default['chef-backup']['s3sync']['directory'] = 'chef'
+default['chef-backup']['s3sync']['cron_hour'] = '1'
+default['chef-backup']['s3sync']['cron_minute'] = '15'
+default['chef-backup']['s3sync']['cron_weekday'] = '*'