+++ metadata.rb
@@ -4,4 +4,6 @@ maintainer_email 'roshan@opsline.com'
 license          'All rights reserved'
 description      'Installs/Configures chef server backup and restore scripts'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.4.0'
+version          '0.5.0'
+
+depends 'logrotate'